% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences_mentxr.R
\name{preferences_mentxr}
\alias{preferences_mentxr}
\title{Assign probabilities based on the preferences of the mentxr}
\usage{
preferences_mentxr(
  info_mentxr,
  data_mentorandx = data_mentorandx,
  p = c(0.1, 0.45, 0.45)
)
}
\arguments{
\item{info_mentxr}{a vector containing four elements: email, nodo, sexo_mentorandx and cat_mentorandx.}

\item{data_mentorandx}{data with all the information of the mentorandxs}
}
\value{
a probabilities vector for each mentorandx of the dataset, based on the preferences of the mentxr
}
\description{
`preferences_mentxr` Main function to assign probabilities based on the preferences of the mentxr
}
\author{
Joaquín Martínez-Minaya <\email{jomartinez@bcamath.org}>
}
